+++ CHANGELOG.md
@@ -179,3 +179,6 @@
 * Add `provider` module that actually implements cryptography!
 * Add `COSE_Key` header to `CoseHeader`, defined in OpenID for Verifiable Credential Issuance draft 13
 * Fix serialization of COSE signature structures
+* Refactor `JsonWebKey`:
+  * Remove `identifier`, please use `keyId` or `jwkThumbprint` directly
+  * Add `equalsCryptographically()` to compare two keys by their cryptographic properties only